refactor: Rename classes in src.db.utils
